ts: extract the shared FNV-1a seed hashing into fnv1a.ts
centerImpl and spreadImpl each carried an identical copy of the FNV-1a
deriveSeed/float64ToBits helpers. Move them to a single fnv1a module and
import from both, removing the duplication.

Acked-by: Andrey Akinshin <andrey.akinshin@gmail.com>
